Enable user input on folder input type and hook up path validation
